+++ ElvUI_MerathilisUI/Core/Changelog/6.31.lua
@@ -1,8 +1,19 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[631] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	RELEASE_DATE = "16.08.2024",
+	FIXES = {
+		"[Mail]: Fix right click error on the characters",
+		"[Archaeology Skin]: Fix the skin",
+	},
+	NEW = {
+		"[Myslot]: Added a skin for it",
+		"[Styling]: It needs more work tho, but i added an option back to enable/disabe the Styling on its own again",
+		"           Also i hook now directly to ElvUI's SetTemplate. Lemme know if you found something weird with it",
+	},
+	IMPROVEMENTS = {
+		"[Event Tracker]: Some improvements to Radiant Echos",
+		"[Game Menu]: Small visual improvements",
+		"[mMediaTags]: Enable also cast icon for target portrait. Thx Blinkii <3",
+	},
 }