+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ### Changes ###
 
-  * [RaidUtility]: Fix part 2
+  * [Mail]: Fix right click error on the characters
+  * [Archaeology Skin]: Fix the skin
+  * [Myslot]: Added a skin for it
+  * [Styling]: It needs more work tho, but i added an option back to enable/disabe the Styling on its own again. Also i hook now directly to ElvUI's SetTemplate. Lemme know if you found something weird with it
+  * [Event Tracker]: Some improvements to Radiant Echos
+  * [Game Menu]: Small visual improvements
+  * [mMediaTags]: Enable also cast icon for target portrait. Thx Blinkii <3